requirements.txt
streamlit
transformers
Pillow
torch
langchain-community
pyttsx3
requests
regex